requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djoser==2.2.2
gunicorn==20.1.0
celery==5.3.1
Pillow==10.1.0
flower==2.0.1